Move desired/scope to scope
