+++ exp/topology/scope/scope_test.go
