+++ exp/topology/scope/state.go
