+++ internal/controllers/topology/cluster/blueprint.go
@@ -22,7 +22,7 @@ import (
 	"github.com/pkg/errors"
 
 	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
-	"sigs.k8s.io/cluster-api/exp/topology/desiredstate/scope"
+	"sigs.k8s.io/cluster-api/exp/topology/scope"
 	tlog "sigs.k8s.io/cluster-api/internal/log"
 )
